:truck: rename leancloud.cloudfunc to leancloud.cloud
